Split camera matrix; added experimental mode to main; changes to arcball control.
Added an experimental flag to main that disables the blit process for
the camera, leaving the background black.  This can be set to True to
enable this functionality or False to return to the camera background.
The camera matrix is split between view and projection for the arcball
control once again.
The arcball has been partially reimplemented, but it's buggy.